+++ b2sdk/raw_api.py
@@ -17,6 +17,7 @@ from typing import Any, Dict, Optional
 
 from .exception import FileOrBucketNotFound, ResourceNotFound, UnusableFileName, InvalidMetadataDirective, WrongEncryptionModeForBucketDefault, AccessDenied, SSECKeyError, RetentionWriteError
 from .encryption.setting import EncryptionMode, EncryptionSetting
+from .replication.setting import ReplicationConfiguration
 from .file_lock import BucketRetentionSetting, FileRetentionSetting, LegalHold
 from .utils import b2_url_encode
 from b2sdk.http_constants import FILE_INFO_HEADER_PREFIX
@@ -126,6 +127,7 @@ class AbstractRawApi(metaclass=ABCMeta):
         lifecycle_rules=None,
         default_server_side_encryption: Optional[EncryptionSetting] = None,
         is_file_lock_enabled: Optional[bool] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         pass
 
@@ -283,6 +285,7 @@ class AbstractRawApi(metaclass=ABCMeta):
         if_revision_is=None,
         default_server_side_encryption: Optional[EncryptionSetting] = None,
         default_retention: Optional[BucketRetentionSetting] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         pass
 
@@ -391,6 +394,7 @@ class B2RawHTTPApi(AbstractRawApi):
         lifecycle_rules=None,
         default_server_side_encryption: Optional[EncryptionSetting] = None,
         is_file_lock_enabled: Optional[bool] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         kwargs = dict(
             accountId=account_id,
@@ -410,6 +414,8 @@ class B2RawHTTPApi(AbstractRawApi):
                   ] = default_server_side_encryption.serialize_to_json_for_request()
         if is_file_lock_enabled is not None:
             kwargs['fileLockEnabled'] = is_file_lock_enabled
+        if replication is not None:
+            kwargs['replicationConfiguration'] = replication.serialize_to_json_for_request()
         return self._post_json(
             api_url,
             'b2_create_bucket',
@@ -695,6 +701,7 @@ class B2RawHTTPApi(AbstractRawApi):
         if_revision_is=None,
         default_server_side_encryption: Optional[EncryptionSetting] = None,
         default_retention: Optional[BucketRetentionSetting] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         assert bucket_info is not None or bucket_type is not None
 
@@ -716,6 +723,8 @@ class B2RawHTTPApi(AbstractRawApi):
                   ] = default_server_side_encryption.serialize_to_json_for_request()
         if default_retention is not None:
             kwargs['defaultRetention'] = default_retention.serialize_to_json_for_request()
+        if replication is not None:
+            kwargs['replicationConfiguration'] = replication.serialize_to_json_for_request()
 
         return self._post_json(
             api_url,