+++ b2sdk/raw_simulator.py
@@ -8,16 +8,19 @@
 #
 ######################################################################
 
-from typing import Optional
 import collections
 import io
 import logging
 import random
 import re
-import time
 import threading
-from contextlib import contextmanager
+import time
 
+from contextlib import contextmanager
+from typing import Optional
+
+from b2sdk.http_constants import FILE_INFO_HEADER_PREFIX, HEX_DIGITS_AT_END
+from b2sdk.replication.setting import ReplicationConfiguration
 from requests.structures import CaseInsensitiveDict
 
 from .b2http import ResponseContextManager
@@ -37,21 +40,21 @@ from .exception import (
     MissingPart,
     NonExistentBucket,
     PartSha1Mismatch,
+    SSECKeyError,
     Unauthorized,
     UnsatisfiableRange,
-    SSECKeyError,
 )
-from .file_lock import BucketRetentionSetting, FileRetentionSetting, NO_RETENTION_BUCKET_SETTING, LegalHold
-from .raw_api import AbstractRawApi, MetadataDirectiveMode, ALL_CAPABILITIES
-from .utils import (
-    b2_url_decode,
-    b2_url_encode,
-    ConcurrentUsedAuthTokenGuard,
-    hex_sha1_of_bytes,
+from .file_lock import (
+    NO_RETENTION_BUCKET_SETTING,
+    BucketRetentionSetting,
+    FileRetentionSetting,
+    LegalHold,
 )
-from b2sdk.http_constants import FILE_INFO_HEADER_PREFIX, HEX_DIGITS_AT_END
-from .stream.hashing import StreamWithHash
 from .file_version import UNVERIFIED_CHECKSUM_PREFIX
+from .raw_api import ALL_CAPABILITIES, AbstractRawApi, MetadataDirectiveMode
+from .stream.hashing import StreamWithHash
+from .utils import ConcurrentUsedAuthTokenGuard, b2_url_decode, b2_url_encode, hex_sha1_of_bytes
+
 
 logger = logging.getLogger(__name__)
 
@@ -492,6 +495,7 @@ class BucketSimulator:
         options_set=None,
         default_server_side_encryption=None,
         is_file_lock_enabled: Optional[bool] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         assert bucket_type in ['allPrivate', 'allPublic']
         self.api = api
@@ -516,6 +520,7 @@ class BucketSimulator:
         self.default_server_side_encryption = default_server_side_encryption
         self.is_file_lock_enabled = is_file_lock_enabled
         self.default_retention = NO_RETENTION_BUCKET_SETTING
+        self.replication = replication
 
     def is_allowed_to_read_bucket_encryption_setting(self, account_auth_token):
         return self._check_capability(account_auth_token, 'readBucketEncryption')
@@ -559,6 +564,9 @@ class BucketSimulator:
             }  # yapf: disable
         else:
             file_lock_configuration = {'isClientAuthorizedToRead': False, 'value': None}
+
+        replication = self.replication and self.replication.as_dict()
+
         return dict(
             accountId=self.account_id,
             bucketName=self.bucket_name,
@@ -571,6 +579,7 @@ class BucketSimulator:
             revision=self.revision,
             defaultServerSideEncryption=default_sse,
             fileLockConfiguration=file_lock_configuration,
+            replicationConfiguration=replication,
         )
 
     def cancel_large_file(self, file_id):
@@ -925,6 +934,7 @@ class BucketSimulator:
         if_revision_is: Optional[int] = None,
         default_server_side_encryption: Optional[EncryptionSetting] = None,
         default_retention: Optional[BucketRetentionSetting] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         if if_revision_is is not None and self.revision != if_revision_is:
             raise Conflict()
@@ -942,6 +952,7 @@ class BucketSimulator:
         if default_retention:
             self.default_retention = default_retention
         self.revision += 1
+        self.replication = replication
         return self.bucket_dict(self.api.current_token)
 
     def upload_file(
@@ -1211,6 +1222,7 @@ class RawSimulator(AbstractRawApi):
         lifecycle_rules=None,
         default_server_side_encryption: Optional[EncryptionSetting] = None,
         is_file_lock_enabled: Optional[bool] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         if not re.match(r'^[-a-zA-Z0-9]*$', bucket_name):
             raise BadJson('illegal bucket name: ' + bucket_name)
@@ -1230,6 +1242,7 @@ class RawSimulator(AbstractRawApi):
             # watch out for options!
             default_server_side_encryption=default_server_side_encryption,
             is_file_lock_enabled=is_file_lock_enabled,
+            replication=replication,
         )
         self.bucket_name_to_bucket[bucket_name] = bucket
         self.bucket_id_to_bucket[bucket_id] = bucket
@@ -1688,6 +1701,7 @@ class RawSimulator(AbstractRawApi):
         if_revision_is=None,
         default_server_side_encryption: Optional[EncryptionSetting] = None,
         default_retention: Optional[BucketRetentionSetting] = None,
+        replication: Optional[ReplicationConfiguration] = None,
     ):
         assert bucket_type or bucket_info or cors_rules or lifecycle_rules or default_server_side_encryption
         bucket = self._get_bucket_by_id(bucket_id)
@@ -1700,6 +1714,7 @@ class RawSimulator(AbstractRawApi):
             if_revision_is=if_revision_is,
             default_server_side_encryption=default_server_side_encryption,
             default_retention=default_retention,
+            replication=replication,
         )
 
     def upload_file(